LayoutTests/platform/chromium/fast/forms/color/color-suggestion-picker-appearance.html

<!DOCTYPE html>
<body>
<input type=color id=color value="#ff0000" list=gray focus>
<datalist id=gray>
    <option>#ffffff</option>
    <option>#eeeeee</option>
    <option>#dddddd</option>
    <option>#cccccc</option>
    <option>#bbbbbb</option>
    <option>#aaaaaa</option>
    <option>#999999</option>
    <option>#888888</option>
    <option>#777777</option>
    <option>#666666</option>
    <option>#555555</option>
    <option>#444444</option>
    <option>#333333</option>
    <option>#222222</option>
    <option>#111111</option>
    <option>#000000</option>
    <option>#F00</option> <!--invalid-->
    <option>red</option> <!--invalid-->
</datalist>
<script>
console.log('Check if window.pagePopupController exists for a page-popup window.');
testRunner.dumpAsText(true);
testRunner.waitUntilDone();
if (window.internals)
    internals.settings.setEnableMockPagePopup(true);

document.getElementById('color').focus();
if (window.eventSender)
    eventSender.keyDown(' ');
var pickerWindow = document.getElementById('mock-page-popup').contentWindow;
pickerWindow.onresize = function() {
    console.log(pickerWindow.pagePopupController);
    testRunner.notifyDone();
}
</script>
</body>
