LayoutTests/platform/chromium/fast/forms/color/color-suggestion-picker-appearance-expected.txt

CONSOLE MESSAGE: line 25: Check if window.pagePopupController exists for a page-popup window.
CONSOLE MESSAGE: line 36: [object PagePopupController]
 
